+++ src/connection.hpp
@@ -13,7 +13,7 @@ namespace yunying {
         std::string received_raw_;
         std::string received_buffer_;
         std::string send_raw_;
-        int send_offset_;
+        size_t send_offset_;
         bool recv_done_;
         bool send_started_;
         bool send_done_;
